fix(lint): basic lint issues for angular
